* upstream:
  Import upstream snapshot from SVN r2670

+++ cmd-capture-pane.c
@@ -1,4 +1,4 @@
-/* $Id: cmd-capture-pane.c 2595 2011-09-21 16:34:34Z tcunha $ */
+/* $Id: cmd-capture-pane.c 2651 2011-12-30 14:16:44Z tcunha $ */
 
 /*
  * Copyright (c) 2009 Jonathan Alvarado <radobobo@users.sourceforge.net>
@@ -59,7 +59,7 @@ cmd_capture_pane_exec(struct cmd *self, struct cmd_ctx *ctx)
 	buf = NULL;
 	len = 0;
 
-	n = args_strtonum(args, 'S', SHRT_MIN, SHRT_MAX, &cause);
+	n = args_strtonum(args, 'S', INT_MIN, SHRT_MAX, &cause);
 	if (cause != NULL) {
 		top = gd->hsize;
 		xfree(cause);
@@ -70,7 +70,7 @@ cmd_capture_pane_exec(struct cmd *self, struct cmd_ctx *ctx)
 	if (top > gd->hsize + gd->sy - 1)
 		top = gd->hsize + gd->sy - 1;
 
-	n = args_strtonum(args, 'E', SHRT_MIN, SHRT_MAX, &cause);
+	n = args_strtonum(args, 'E', INT_MIN, SHRT_MAX, &cause);
 	if (cause != NULL) {
 		bottom = gd->hsize + gd->sy - 1;
 		xfree(cause);
